requirements.txt
openai-whisper
gradio
python-docx
transformers
deep-translator
gtts
langdetect
httpx
torch
